[FIX] models: do not check Python constraints on partial updates
Consider a model with a constraint method that depends on fields `currency_id`
(stored) and `country_id` (computed):

    @api.constrains('country_id', 'currency_id')
    def _check_country_currency(self):
        # check whether country_id and currency_id match each other

When writing on both `country_id` and `currency_id`, the method is invoked
twice: once when `currency_id` is updated, and again when both fields are
updated.  There are cases where the first invocation fails, but overall is
should not fail.

Fix this by not checking constraints on partial updates.

OPW 1817493