wip core: add an exists / not exists internal operator
And use it for *2m =/!= False queries, to avoid carrying potentially
millions of ids back and forth and ending up with postgres throwing an
`ERROR: invalid memory alloc request`.

One thing this doesn't quite solve is if the leaf is not at the root
of the query e.g. `a_id.b_ids = False` should be slightly less awful
but it's still not going to be great as the m2o will execute the
search on the subfield in full and still result in an `(id in <ids>)`
leaf.