[IMP] website_form: made the form data dynamic
Previously, once we designed the form and saved the page, the form
was not changed even if there were changes in the db related to added
fields (for example, in case of m2o field, a newly created record was
never appeared on the selection).

This commit re-fetches the values from database while landing
on the page having forms and updates the values for such fields
to keep it in sync with the database values.

TaskID - 1874350