[IMP] web: SearchPanel on mobile
SearchPanel widget in kanban views has been introduced in commit
https://github.com/odoo/odoo/pull/30517/commits/25c9355c02485e85914430188f3281246fc54cda.

It allows to refine search according to specific dimensions. This widget
is displayed as a sidebar to the left of the kanban view. But it's also
missing a representation for small screens, and then disabled on mobile.

This commit enables it on mobile giving the same features as the desktop
version but displayed as an additional row at the top of the kanban
records.

It displays a summary of the currently selected filters/categories.
Clicking on it opens a modal screen presenting the filters and/or
categories similarly to the desktop version.

Task ID: 1917277