[IMP] stock: forcast report improvements
The forecast report added in b311d9c always shows the stock in all locations
even in multicompany/multiwarehouse.
This commit change the SQL query that gather the product quantities over
time to take the location into account and be able to filter on it.

regarding stock.move, the location is the source one in cast of outgoing
move and the destination one in case of the ingoing move.

Furthermore, the SQL query doesn't filter the product on their type. That
means a stock move on consumable product will be shown even if this product

Task : 2003485