[IMP] website_form: known partners  add as followers if in a CC
PURPOSE

People sending tickets through the website might add their vendors or consultant
in the CC field but the vendor or consultant is not added as a follower of the
task.

SPECIFICATIONS

It should be added automatically so they can easily follow what's happening to
their clients. Hence, this commit adds known partners as followers if added in
a CC.

LINKS

PR https://github.com/odoo/odoo/pull/48497
Task 1974030