[IMP] point_of_sale: move formatting methods to the PosModel
Formatting methods (format_currency(), format_currency_no_symbol(),
format_pr()) are currently available in every POS's widgets (through the
BaseWidget class). This limits and complexify their usage in the model
(by requiring a reference to the Chrome instance and/or creating an
empty BaseWidget instance *only* to access those methods).

This commit moves their implementation to the model, removing the need
to such references on UI widgets in the model, but keeping a reference
to those methods in BaseWidget (delegating to the model) for
compatibility reason.