[IMP] base: dump stack in _button_immediate_function
Sometime, test test_01_cow_views_unlink_on_module_update fails
because _button_immediate_function raises an error on
'SELECT * FROM ir_cron FOR UPDATE NOWAIT'. The reason is not obvious
but it may be linked to a concurrent cursor locking cron.
This commits simply add a dumpstack in order to help to debug
this race condition.