[IMP] website_sale(_product_configurator): disable quantity everywhere.
Purpose of the commit is too make the customize option
"Select Quantity" works everywhere.

When we disable it, it's only disabled from the product page. it's
still active on the cart and on the product config modal.

So if the user decide to disable the "Select Quantity", disable it everywhere
(cart, product config modal, product itself)

Task-2008472
Closes: #37951