[FIX] account: no notify on Vendor Bill
* From Accounting dashboard create a new bendor bill, select a customer
and at least a product.
* Save
* Add as follower of the discussion an user (like Marc Demo).
* Edit Marc Demo preferences to receive notifications via Odoo
* Edit (in debug mode) the subscription preferences, selecting all options
* Now validate the invoice

Marc Demo will not receive the notification for the validation, even if
the subscription was active.

This work with Customer Invoices but not with Vendor Bills because when
generating the notification the "subtype_xmlid" is checked and will
return a value only if the invoice is an "outgoing" one.

Releasing the constraint allow to send notifications also for other
types of invoices.

opw-2076532