[IMP] mail, mass_mailing: make technical fields easily understandable
Right now, `reply_to` field on email template is misleading due to
poor explaination on the same.

This commit improves the placeholder and tooltip of the field to make the
purpose of the field more clear, especially for non technical users.

This commit also improves usability for mail composer wizard. Field
`no_auto_thread` along with it's label sounds a bit technical, and so
instead we now have a selection field for the same with more obvious
label and options.

Task ID: 2117639
PR #40931