[IMP] web: apply more precedence for mobile searchview css rules
Before this commit, few of the css rules related to mobile searchview
were having less precedence and so UI was not as expected.

For example, the css for the "See results" CTA were not applied
correctly (e.g. `font-size`).

This commit fixes this issue by adding proper precedence for the mobile
searchview related rules and thus avoiding such issues in future.

Task ID: 2152164