[IMP] server.py: Async Worker
[IMP] server.py: Async Worker

We add a new Worker server type whose mission is to continuously process
asynchronous tasks in the background.

The workers block up to one minute waiting for an event to be sent on
the `odoo_async` channel on the meta `postgres` database. Other workers
automatically send and event on that channel every time a job has been
enqueued. See the `IrAsync.call` and `IrAsync._notify_db` methods.

Because sometime events do not effectively wake up the listening workers,
they are automatically woke up every minute to perform an active polling
of the connection.

Async and CRON Worker have a lot of commonality, they both sleep for an
event before processing jobs on several other databases. The differences
are (i) the Async worker may be explicitly woken up by an event whereas
CRON workers "just" sleep one plain minute, (ii) the Async worker receive
a precise list of databases to process whereas the CRON just process.

Because they have much in common, we captured commonalities in an
abstract super class. This ensure all optimizations done for the CRON
workers are used in the new Async worker too.

Task: 2081201