[IMP] hr(_fleet): Display license plate on the employee
Purpose of the task is to find the employee's private vehicle easily
by place number.

HR:
 -License Plate field is added to the Employee Private Address tab so
the employee can be searched through the license plate.
Hr_Fleet:
 - When the driver of the company's vehicle is assigned, write the
vehicle number to the employee and set the vehicle number
blank to the employee of previous driver.
 - If the employee is not linked with the user then find the employee
by the private address and update the vehicle number on that.
 - Another case is when vehicle number is updated on the vehicle then
update that number to the related driver's employee.

TaskID: 2158962
Related: https://github.com/odoo/enterprise/pull/7786
Closes: #43491