[IMP] event: merge the SMS template and the mail template into one field
Purpose
=======
We want to limite the size of the communication tab. For that purpose,
we merge the mail template and the SMS template into one reference field.

Technical
=========
As we can not set a domain on a reference field, we added a context key
and in the `_name_search` of the `mail/sms.template` we filter with the
domain we want if the key exists.

Task 2127615
See odoo/enterprise/pull/7701
See odoo/upgrade/pull/848