[FIX] base: importing abc from collections is deprecated
It was originally slated for removal in 3.8 (and this code would have
broken) but that's apparently been pushed back to 3.9, and might get
pushed further back to 3.10 / 4.

Fixing this in 12 rather than 11 because 11 is still supposed to run
on Python 2, and collections.abc never existed in Python 2 so this
change would require adding a conditional / alternative
import. Simpler to just implement the straightforward fix in 12.