[IMP] project: propagate internal subtypes to tasks
before this commit:
while changing subtypes of the project it was only
propagating those subtypes to the task, whose parent_id is set.
internal subtypes are not propagating in the task of a project.

So, filter internal subtypes and concat them with the subtypes which
have parent_id.

after this commit:
internal subtypes will also propagate to their tasks

LINKS:
PR https://github.com/odoo/odoo/pull/47336
task 2205643