[FIX] account: use the date when reversing
When unreconciling, if an 'account.full.reconcile' has to be unlinked
the 'reverse_moves' will be called without date, defaulting to today.
This should not be the case becase during the period between the
original date and "today", the account balances used in the exchange
difference will not be accurate.

Adding the move date to the call fix the issue

opw-2209675