[REM] *: a bunch of metaclasses
Initial goal was to remove the intermediate classes used to handle
cross-version metaclassing (class attribute on P2 versus kwarg on P3).

However Python 3.6 introduced __init_subclass__, which is invoked
right after a subclass of the class defining the method is
created. This means the common cases of using metaclasses for basic
post-creation registration of transformation can now be performed by
just defining this one method.

Since the conversion removes a fair amount of code, it seems like a
good idea.

* gcal metaclass seems completely useless, the kw feature is never
  used and the fallback on class attributes works OOTB, I'm not even
  sure the __getitem__ is necessary
* drivermeta is trivial registration
* command is trivial configuration & registration

Field, Controller and BaseModel only remove the dummy intermediate
class.