[FIX] l10n_fr_pos_cert: Allow value between 0 and 1.
In certain context, we cannot decrease the quantity of new products. But this can lead to issue when we want a value between 0 and 1 because the default value of the quantity is 1.
We allow the user to set it below 1 if the current value is equal to 1. Once it's modified, it's impossible to change it anymore.

task-id: 2230922