[FIX] mail: Do not delete entities on messaging service destroy
Ideally, we would like to remove all entities on destroy.
However, some components that should have been destroyed may still
being re-rendered by OWL and access core entities that no longer
exist, such as `messaging`.

We could have guarded every uses of these entities in code of
components, but this is too many changes and pollute overall code
due to essential architectural flaw of observable in OWL.

Idea is to re-introduce this cleanup after fixing the issue on
observable / components in regard to store management with entities.

Still, it propagates `stop()` on destroying messaging service, so
that event listeners are removed at least.

This commit also fixes an issue in which device was not auto-updating
itself when resizing window.

Task-Id 2233981