[IMP] base, web: custom title in error modal
In a ancient age were the sole exception were `except_osv` and
`except_orm`, every exception needed two arguments, an error title and
an error message. With the introduction of dedicated exceptions in  only
the error message remained as the title was given by the exception type.

It turns out most of the time the generic `UserError` exception is
raised for missuses that don't fall in any other category and that the
error modal the user see is entitled "Something went wrong" which is too
much generic and unnecessary scary.

We introduced a way to customize the modal title by providing a `title`
argument when raising a odoo custom exception.

Task: 2206851