[FIX] web, mail: make mail links work through oauth login
For our oauth flow, we bundle the URL we're ultimately
reopening (after the user has authorized the application and this
authorization has been validated) into the `state` parameter (`r`).

This URL is generated by the server, which means it can't access
the *fragment*, which means while the fragment is kept from
`/web#action=xxx` to `/web/login` it can't be merged into the
"terminal" URL, and thus gets lost when we go through the
authorization server.

This change implements a relatively simple fix which should solve most
of the issue without large changes or horrible hacks: mail already has
an indirection `/mail/view` which uses query string and derefs that to
whatever action description is necessary. It's what used for e.g. the
"view task" link in notification emails.

The issue with it, is that it ends up deref'ing to a fragment-based
URL. Instead, bounce to /web with a `redirect` query string parameter
of that fragment-based URL: auth_oauth will use that parameter as its
final URL instead of generating one from scratch, and if the user is
already logged in `/web` will just redirect to the "correct" `/web`
invocation.

Then fix /web so it doesn't lose the `redirect` parameter when
redirecting non-logged users to the login page: `/web` will use that
parameter if the user is already logged, and `/web/login` will *also*
use it on login (including auth_oauth), so it's just that one little
piece which is missing.