[FIX] mail: inconsistent attachment download on mobile
In the DocumentViewer, when the user attempts to download a file on iOS,
the action is not recognized as a download but only as a redirection,
leaving the user stuck with the document opened but no way to get back
to the app.

This commit fixes it by using the `session.get_file()` (which is the
standard way of downloading a file in Odoo) instead of a redirection
through `window.location`. This has the nice benefit of being more
consistent not only on iOS but also on Android for files' download.

opw-2260101