[FIX] partner_autocomplete: fix dropdown style to make avatar clickable
Previolsy, in company form view avatar was not clickable in small devices
due to a `block` div was overlapping the avatar widget and preveting
click on it.

This commit fixes the issue by preventing the overlap and thus making
the image widget on company form clickable.

Note : A proper fix would be to  move `partner_gid` field somewhere
else in the comapny form, which comes between `oe_avatar` and `oe_title`
and messes with the layout in small devices. However, considering
stable version, changing fields posistion should be avoided.

Task Id: 2231950
PR: #53798

Co-authored-by: D J <dja@odoo.com>