-- waiting for forward-port --
[IMP] web_editor, portal: review img/fa display rules mixup

Part of https://github.com/odoo/odoo/pull/53930
task-2157252

[IMP] web_editor: hide we-row with no visible widgets

Part of https://github.com/odoo/odoo/pull/53930
task-2157252

[FIX] web_editor: make sure user events bubble before updating UI

Example where this was a problem: when clicking on a field type for
website form, the whole option is rerendered and we never go through the
"close" method of the clicked we-select element (which will be important
in the future).

Part of https://github.com/odoo/odoo/pull/53930
task-2157252

[IMP] web_editor, *: remove snippet panel block icons

*: website, mass_mailing

Part of https://github.com/odoo/odoo/pull/53930
task-2157252

[IMP] web_editor, *: review the editor design CSS

*: web, mass_mailing, website, website_form, website_sale

New CSS design of the editor (step 1/x).

This work is unfortunately far from perfect but is a first step towards
what we want. In particular, much CSS linked to elements that are meant
to be removed/moved in the upcoming weeks was done in a huge rush but
will thus hopefully be removed/reviewed before the 14.0 release.

As the design is now far more complex including gradients and other
"complex" elements, it was also very difficult to keep a design based on
css-variable so this system was removed, which means that the editor now
appears the same everywhere (backend, frontend, ...). The original goal
of this system was to have an adapted style for backend, frontend and
website visitors. Unfortunately, the current result was a very good
frontend editor design, a not-so-good backend editor design (for mass
mailing at least) and a totally unadapted design for the website
visitors. Making this perfect will be achievable by loading different
assets for those 3 areas and hopefully those assets will only differ by
the values of editor scss variables.

Part of https://github.com/odoo/odoo/pull/53930
task-2157252

[IMP] web_editor: allow to use images in the left panel

Part of https://github.com/odoo/odoo/pull/53930
task-2157252

Co-authored-by: Benjamin Vray <bvr@odoo.com>

[IMP] web_editor, *: introduce new "button group" editor widget

*: website, website_form

Basically work just as a we-select but with a different design.

Part of https://github.com/odoo/odoo/pull/53930
task-2157252

[IMP] web_editor, *: review background color/image options UI

*: website, mass_mailing, website_blog, doc

Part of https://github.com/odoo/odoo/pull/53930
task-2157252

[FIX] website: remove duplicated filter option for blockquote

The blockquote in its "cover" layout allows to add a background with
a dedicated colored filter on top of it. That option was a bit buggy
since the filter option only appeared if the page was saved first. As
we now have a generic option to apply a colored filter on image
themselves, that option became useless anyway.

Part of https://github.com/odoo/odoo/pull/53930
task-2157252

[IMP] website, *: add prettier preview of border styles

*: web_editor

Part of https://github.com/odoo/odoo/pull/53930
task-2157252

[IMP] website: remove animate scrolling option for table of content

Part of https://github.com/odoo/odoo/pull/53930
task-2157252

[IMP] website, *: review options' order/label + add "⌙" markers

*: web_editor

Some of those "⌙" markers will become foldable options in the future.

Part of https://github.com/odoo/odoo/pull/53930
task-2157252